MDED-8 Удален TabBar и Charts VC + fix bugs
