+++ App/MdEditor/Sources/SceneDelegate.swift
@@ -3,7 +3,6 @@
 //
 
 import UIKit
-import TaskManagerPackage
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
